Move project card to projects page
