+++ src/pages/projects.tsx
@@ -1,5 +1,10 @@
 import React from "react"
+import { GatsbyImage } from "gatsby-plugin-image"
+import Card from '@material-ui/core/Card';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
 import Container from "@material-ui/core/Container"
+import Link from "@material-ui/core/Link"
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 
@@ -7,7 +12,42 @@ import Layout from "../layouts/Layout";
 import DefaultFooter from "../components/Footer"
 import SEO from "../components/SEO"
 import { usePages } from "../hooks/use-pages"
-import ProjectCard from "../components/project"
+import { Icon } from "../components/icons"
+import { Page } from "../models"
+
+interface ProjectCardProps {
+  item: Page
+  noImage?: boolean
+}
+
+export const ProjectCard = (props: ProjectCardProps): JSX.Element => {
+  const { slug, title, resume, image, caption, stack } = props.item
+  const stackIcons = stack.split(",")
+  return (
+    <Card component="article">
+      <CardContent>
+        <GatsbyImage
+          as={CardMedia}
+          image={image}
+          alt={caption}
+        />
+        <Typography component="h2" variant="h5">
+          <Link href={`/${slug}`}>{title}</Link>
+        </Typography>
+        <Typography variant="body1" gutterBottom>
+          {resume}
+        </Typography>
+        <Typography>
+          {stackIcons.map(icon => (
+            <span key={icon}>
+              <Icon name={icon} /> {icon}
+            </span>
+          ))}
+        </Typography>
+      </CardContent>
+    </Card>
+  )
+}
 
 const ProjectsPage = (): JSX.Element => {
   const projects = usePages()
